+++ .changeset/unlucky-mails-search.md
@@ -1,0 +1,5 @@
+---
+"jazz-react-native": minor
+---
+
+Change the way the JazzProvider is created and make the hooks available as top-level imports.
