--- examples/chat-rn-clerk/src/jazz.tsx
@@ -1,4 +1,0 @@
-import { createJazzRNApp } from "jazz-react-native";
-
-export const Jazz = createJazzRNApp();
-export const { useAccount, useCoState, useAcceptInvite, kvStore } = Jazz;
