+++ homepage/homepage/app/docs/[framework]/[...slug]/project-setup/react-native.mdx
@@ -214,7 +214,7 @@ Refer to the Jazz + React Native demo projects for implementing authentication:
 
 In the demos, you'll find details on:
 
-- Using Jazz.Provider with your chosen authentication method
+- Using JazzProvider with your chosen authentication method
 - Defining a Jazz schema
 - Creating and subscribing to covalues
 - Handling invites
