+++ packages/jazz-react-native/package.json
@@ -16,12 +16,17 @@
       "react-native": "./dist/crypto/index.js",
       "types": "./dist/crypto/index.d.ts",
       "default": "./dist/crypto/index.js"
+    },
+    "./testing": {
+      "react-native": "./dist/testing.js",
+      "types": "./dist/testing.d.ts",
+      "default": "./dist/testing.js"
     }
   },
   "license": "MIT",
   "dependencies": {
     "@scure/base": "1.2.1",
-    "@scure/bip39": "^1.3.0",
+    "jazz-react-core": "workspace:*",
     "cojson": "workspace:*",
     "cojson-transport-ws": "workspace:*",
     "jazz-tools": "workspace:*",
