--- tests/e2e-react-native/src/jazz.tsx
@@ -1,4 +1,0 @@
-import { createJazzRNApp } from "jazz-react-native";
-
-export const Jazz = createJazzRNApp();
-export const { useAccount, useCoState, useAcceptInvite } = Jazz;
