+++ tests/e2e-react-native/src/screens/SimpleSharing.tsx
@@ -1,9 +1,9 @@
 import * as Clipboard from "expo-clipboard";
 import { createInviteLink } from "jazz-react-native";
+import { useAccount, useCoState } from "jazz-react-native";
 import { Group, ID } from "jazz-tools";
 import { useState } from "react";
 import { Alert, Button, Text, View } from "react-native";
-import { useAccount, useCoState } from "../jazz";
 import { CoMapWithText } from "../schema";
 
 export function SimpleSharing() {
